i3/requirements.txt
musicbrainzngs==0.6
requests==2.22.0
screeninfo==0.3.1
Pillow==6.0.0
